+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changelog
 =========
 
+0.0.2 (2023-11-12)
+------------------
+
+- title and description used in the interface are now customizable via the `--title`
+  and `--description` command-line options
+
+
 0.0.1 (2023-11-03)
 ------------------